apps/ui: split site-dropdown into flat sibling files
The single 616-line site-dropdown/index.tsx mixed a trigger forwardRef,
a row primitive, two full popup views, and the sync-activity indicator
in one place. This breaks it up into sibling modules under the same
folder (no nested subfolders), each with its own colocated
`*.module.css`. The orchestrator `index.tsx` now only owns view state,
hook wiring, and handler composition.

Also drops the no-longer-needed custom tooltip wrapper and its app-root
Tooltip.Provider — @wordpress/ui's IconButton already provides tooltip
coverage where it matters, and the sync indicator renders as an
IconButton so it inherits the same behaviour.
