+++ apps/ui/src/app.tsx
@@ -3,7 +3,6 @@ import { RouterProvider } from '@tanstack/react-router';
 import { defaultI18n } from '@wordpress/i18n';
 import { I18nProvider } from '@wordpress/react-i18n';
 import { privateApis } from '@wordpress/theme';
-import * as Tooltip from '@/components/tooltip';
 import { ConnectorProvider, queryClient } from '@/data/core';
 import { useSyncSitesWithEvents } from '@/data/queries/use-sites';
 import { usePrefersColorScheme } from '@/hooks/use-prefers-color-scheme';
@@ -39,12 +38,7 @@ export function App( { connector }: AppProps ) {
 				<SiteEventsBridge />
 				<I18nProvider i18n={ defaultI18n }>
 					<ThemeProvider isRoot color={ themeColor } density="compact">
-						{ /* A shared provider keeps the tooltip delay consistent across
-							every consumer (including @wordpress/ui's IconButton, which
-							wraps its own Provider with delay=0 — those nest fine). */ }
-						<Tooltip.Provider delay={ 250 }>
-							<RouterProvider router={ router } />
-						</Tooltip.Provider>
+						<RouterProvider router={ router } />
 					</ThemeProvider>
 				</I18nProvider>
 			</QueryClientProvider>
