+++ apps/ui/src/components/site-dropdown/dropdown-trigger.module.css
@@ -1,0 +1,64 @@
+.trigger {
+	gap: var(--wpds-dimension-padding-lg);
+}
+
+.site {
+	font-weight: 600;
+}
+
+.status {
+	display: inline-flex;
+	align-items: center;
+	gap: var(--wpds-dimension-padding-xs);
+}
+
+.dot {
+	width: 6px;
+	height: 6px;
+	border-radius: 50%;
+	flex-shrink: 0;
+}
+
+/* Hardcoded because `--wpds-color-fg-content-success` is #002900 (reads as
+   black at 6px) and the critical/accent equivalents don't exist. */
+.dot_running {
+	background-color: #16a34a;
+}
+
+.dot_stopped {
+	background-color: #dc2626;
+}
+
+.dot_transitioning {
+	background-color: #2563eb;
+	animation: siteDotBlink 1s ease-in-out infinite;
+}
+
+.dot_live {
+	background-color: #2563eb;
+}
+
+@keyframes siteDotBlink {
+	0%,
+	100% {
+		opacity: 1;
+	}
+	50% {
+		opacity: 0.25;
+	}
+}
+
+.env {
+	color: var(--wpds-color-fg-content-neutral-weak);
+	font-size: var(--wpds-font-size-sm);
+}
+
+.url {
+	color: var(--wpds-color-fg-content-neutral-weak);
+	opacity: 0.6;
+	font-size: var(--wpds-font-size-sm);
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	min-width: 0;
+}
