+++ apps/ui/src/components/site-dropdown/main-view.module.css
@@ -1,0 +1,22 @@
+.rows {
+	display: flex;
+	flex-direction: column;
+	padding: var(--wpds-dimension-padding-sm) 0;
+}
+
+.rowActions {
+	display: flex;
+	align-items: center;
+	gap: var(--wpds-dimension-padding-xs);
+}
+
+.footer {
+	display: flex;
+	gap: var(--wpds-dimension-padding-sm);
+	padding: 0 var(--wpds-dimension-padding-lg) var(--wpds-dimension-padding-md);
+}
+
+.footerButton {
+	flex: 1;
+	justify-content: center;
+}
