+++ apps/ui/src/components/site-dropdown/popover-row.module.css
@@ -1,0 +1,37 @@
+.row {
+	display: flex;
+	align-items: center;
+	gap: var(--wpds-dimension-padding-md);
+	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-lg);
+	min-height: 44px;
+}
+
+.text {
+	flex: 1;
+	min-width: 0;
+	display: flex;
+	flex-direction: column;
+}
+
+.label {
+	font-size: var(--wpds-font-size-sm);
+	line-height: 1.2;
+	font-weight: 500;
+	color: var(--wpds-color-fg-content-neutral);
+}
+
+.sublabel {
+	margin-top: 4px;
+	font-size: var(--wpds-font-size-sm);
+	line-height: 1.2;
+	color: var(--wpds-color-fg-content-neutral-weak);
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+}
+
+.action {
+	flex-shrink: 0;
+	color: var(--wpds-color-fg-content-accent);
+	font-size: var(--wpds-font-size-sm);
+}
