+++ apps/ui/src/components/site-dropdown/publish-picker-view.module.css
@@ -1,0 +1,95 @@
+.picker {
+	display: flex;
+	flex-direction: column;
+	min-height: 160px;
+}
+
+.header {
+	display: flex;
+	align-items: center;
+	gap: var(--wpds-dimension-padding-sm);
+	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-md);
+	border-bottom: 1px solid var(--wpds-color-stroke-neutral-weak);
+}
+
+.title {
+	font-size: var(--wpds-font-size-sm);
+	font-weight: 500;
+	color: var(--wpds-color-fg-content-neutral);
+}
+
+.body {
+	max-height: 240px;
+	overflow-y: auto;
+}
+
+.status {
+	padding: var(--wpds-dimension-padding-lg);
+	color: var(--wpds-color-fg-content-neutral-weak);
+	font-size: var(--wpds-font-size-sm);
+	text-align: center;
+}
+
+.list {
+	list-style: none;
+	margin: 0;
+	padding: var(--wpds-dimension-padding-xs) 0;
+}
+
+.item {
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	gap: 2px;
+	width: 100%;
+	background: transparent;
+	border: none;
+	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-lg);
+	cursor: pointer;
+	font: inherit;
+	color: inherit;
+	text-align: left;
+}
+
+.item:hover,
+.item:focus-visible {
+	background-color: var(--wpds-color-bg-surface-neutral-hover);
+	outline: none;
+}
+
+.itemName {
+	font-size: var(--wpds-font-size-sm);
+	font-weight: 500;
+	color: var(--wpds-color-fg-content-neutral);
+}
+
+.itemUrl {
+	font-size: var(--wpds-font-size-sm);
+	color: var(--wpds-color-fg-content-neutral-weak);
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	max-width: 100%;
+}
+
+.create {
+	display: flex;
+	align-items: center;
+	gap: var(--wpds-dimension-padding-sm);
+	width: 100%;
+	background: transparent;
+	border: none;
+	border-top: 1px solid var(--wpds-color-stroke-neutral-weak);
+	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-lg);
+	cursor: pointer;
+	font: inherit;
+	font-size: var(--wpds-font-size-sm);
+	color: var(--wpds-color-fg-content-accent);
+	text-align: left;
+}
+
+.create:hover,
+.create:focus-visible {
+	background-color: var(--wpds-color-bg-surface-neutral-hover);
+	outline: none;
+}
