+++ apps/ui/src/components/site-dropdown/style.module.css
@@ -5,141 +5,6 @@
 	min-width: 0;
 }
 
-.trigger {
-	gap: var(--wpds-dimension-padding-lg);
-}
-
-/* The indicator renders as an IconButton so we inherit its tooltip wiring.
-   Override the default button background/color with the state-specific hue.
-   The `!important` keeps the color stable across @wordpress/ui's hover and
-   disabled states — a disabled IconButton otherwise washes out the fill. */
-.syncIndicator,
-.syncIndicator:hover,
-.syncIndicator:focus-visible,
-.syncIndicator[disabled] {
-	color: #ffffff !important;
-	border-color: transparent !important;
-	border-radius: 999px;
-	opacity: 1;
-}
-
-/* Blink while in progress. Hardcoded hues mirror .triggerDot_transitioning so
-   the indicator visually matches the rest of the dropdown state colors. */
-.syncIndicator_pending,
-.syncIndicator_pending:hover,
-.syncIndicator_pending[disabled] {
-	background-color: #2563eb !important;
-	animation: siteSyncBlink 1.1s ease-in-out infinite;
-}
-
-.syncIndicator_success,
-.syncIndicator_success:hover,
-.syncIndicator_success[disabled] {
-	background-color: #16a34a !important;
-}
-
-.syncIndicator_error,
-.syncIndicator_error:hover,
-.syncIndicator_error:focus-visible {
-	background-color: #dc2626 !important;
-}
-
-@keyframes siteSyncBlink {
-	0%,
-	100% {
-		opacity: 1;
-	}
-	50% {
-		opacity: 0.35;
-	}
-}
-
-.errorPopup {
-	max-width: 320px;
-	padding: var(--wpds-dimension-padding-md);
-	background: var(--wpds-color-bg-surface-strong, #ffffff);
-	border: 1px solid var(--wpds-color-stroke-neutral-weak);
-	border-radius: var(--wpds-dimension-radius-md);
-	box-shadow: var(--wpds-shadow-elevation-md);
-	color: var(--wpds-color-fg-content-neutral);
-	font-size: var(--wpds-font-size-sm);
-	line-height: 1.4;
-	display: flex;
-	flex-direction: column;
-	gap: var(--wpds-dimension-padding-xs);
-}
-
-.errorPopupTitle {
-	font-weight: 600;
-	color: #dc2626;
-}
-
-.errorPopupMessage {
-	white-space: pre-wrap;
-	word-break: break-word;
-}
-
-.triggerSite {
-	font-weight: 600;
-}
-
-.triggerStatus {
-	display: inline-flex;
-	align-items: center;
-	gap: var(--wpds-dimension-padding-xs);
-}
-
-.triggerDot {
-	width: 6px;
-	height: 6px;
-	border-radius: 50%;
-	flex-shrink: 0;
-}
-
-/* Hardcoded because `--wpds-color-fg-content-success` is #002900 (reads as
-   black at 6px) and the critical/accent equivalents don't exist. */
-.triggerDot_running {
-	background-color: #16a34a;
-}
-
-.triggerDot_stopped {
-	background-color: #dc2626;
-}
-
-.triggerDot_transitioning {
-	background-color: #2563eb;
-	animation: siteDotBlink 1s ease-in-out infinite;
-}
-
-.triggerDot_live {
-	background-color: #2563eb;
-}
-
-@keyframes siteDotBlink {
-	0%,
-	100% {
-		opacity: 1;
-	}
-	50% {
-		opacity: 0.25;
-	}
-}
-
-.triggerEnv {
-	color: var(--wpds-color-fg-content-neutral-weak);
-	font-size: var(--wpds-font-size-sm);
-}
-
-.triggerUrl {
-	color: var(--wpds-color-fg-content-neutral-weak);
-	opacity: 0.6;
-	font-size: var(--wpds-font-size-sm);
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	min-width: 0;
-}
-
 .popup {
 	min-width: 320px;
 	padding: 0;
@@ -152,161 +17,4 @@
 .popup button:focus:not(:focus-visible) {
 	background-color: var(--wp-ui-button-background-color);
 	border-color: var(--wp-ui-button-border-color);
-}
-
-.rows {
-	display: flex;
-	flex-direction: column;
-	padding: var(--wpds-dimension-padding-sm) 0;
-}
-
-.row {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-padding-md);
-	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-lg);
-	min-height: 44px;
-}
-
-.rowText {
-	flex: 1;
-	min-width: 0;
-	display: flex;
-	flex-direction: column;
-}
-
-.rowLabel {
-	font-size: var(--wpds-font-size-sm);
-	line-height: 1.2;
-	font-weight: 500;
-	color: var(--wpds-color-fg-content-neutral);
-}
-
-.rowSublabel {
-	margin-top: 4px;
-	font-size: var(--wpds-font-size-sm);
-	line-height: 1.2;
-	color: var(--wpds-color-fg-content-neutral-weak);
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.rowAction {
-	flex-shrink: 0;
-	color: var(--wpds-color-fg-content-accent);
-	font-size: var(--wpds-font-size-sm);
-}
-
-.localActions {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-padding-xs);
-}
-
-.footer {
-	display: flex;
-	gap: var(--wpds-dimension-padding-sm);
-	padding: 0 var(--wpds-dimension-padding-lg) var(--wpds-dimension-padding-md);
-}
-
-.footerButton {
-	flex: 1;
-	justify-content: center;
-}
-
-.picker {
-	display: flex;
-	flex-direction: column;
-	min-height: 160px;
-}
-
-.pickerHeader {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-padding-sm);
-	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-md);
-	border-bottom: 1px solid var(--wpds-color-stroke-neutral-weak);
-}
-
-.pickerTitle {
-	font-size: var(--wpds-font-size-sm);
-	font-weight: 500;
-	color: var(--wpds-color-fg-content-neutral);
-}
-
-.pickerBody {
-	max-height: 240px;
-	overflow-y: auto;
-}
-
-.pickerStatus {
-	padding: var(--wpds-dimension-padding-lg);
-	color: var(--wpds-color-fg-content-neutral-weak);
-	font-size: var(--wpds-font-size-sm);
-	text-align: center;
-}
-
-.pickerList {
-	list-style: none;
-	margin: 0;
-	padding: var(--wpds-dimension-padding-xs) 0;
-}
-
-.pickerItem {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-	gap: 2px;
-	width: 100%;
-	background: transparent;
-	border: none;
-	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-lg);
-	cursor: pointer;
-	font: inherit;
-	color: inherit;
-	text-align: left;
-}
-
-.pickerItem:hover,
-.pickerItem:focus-visible {
-	background-color: var(--wpds-color-bg-surface-neutral-hover);
-	outline: none;
-}
-
-.pickerItemName {
-	font-size: var(--wpds-font-size-sm);
-	font-weight: 500;
-	color: var(--wpds-color-fg-content-neutral);
-}
-
-.pickerItemUrl {
-	font-size: var(--wpds-font-size-sm);
-	color: var(--wpds-color-fg-content-neutral-weak);
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	max-width: 100%;
-}
-
-.pickerCreate {
-	display: flex;
-	align-items: center;
-	gap: var(--wpds-dimension-padding-sm);
-	width: 100%;
-	background: transparent;
-	border: none;
-	border-top: 1px solid var(--wpds-color-stroke-neutral-weak);
-	padding: var(--wpds-dimension-padding-sm) var(--wpds-dimension-padding-lg);
-	cursor: pointer;
-	font: inherit;
-	font-size: var(--wpds-font-size-sm);
-	color: var(--wpds-color-fg-content-accent);
-	text-align: left;
-}
-
-.pickerCreate:hover,
-.pickerCreate:focus-visible {
-	background-color: var(--wpds-color-bg-surface-neutral-hover);
-	outline: none;
 }
