+++ apps/ui/src/components/site-dropdown/sync-activity-indicator.module.css
@@ -1,0 +1,69 @@
+/* The indicator renders as an IconButton so we inherit its tooltip wiring.
+   Override the default button background/color with the state-specific hue.
+   The `!important` keeps the color stable across @wordpress/ui's hover and
+   disabled states — a disabled IconButton otherwise washes out the fill. */
+.indicator,
+.indicator:hover,
+.indicator:focus-visible,
+.indicator[disabled] {
+	color: #ffffff !important;
+	border-color: transparent !important;
+	border-radius: 999px;
+	opacity: 1;
+}
+
+/* Blink while in progress. Hardcoded hues mirror the dropdown trigger dot
+   so the indicator visually matches the rest of the dropdown state colors. */
+.pending,
+.pending:hover,
+.pending[disabled] {
+	background-color: #2563eb !important;
+	animation: siteSyncBlink 1.1s ease-in-out infinite;
+}
+
+.success,
+.success:hover,
+.success[disabled] {
+	background-color: #16a34a !important;
+}
+
+.error,
+.error:hover,
+.error:focus-visible {
+	background-color: #dc2626 !important;
+}
+
+@keyframes siteSyncBlink {
+	0%,
+	100% {
+		opacity: 1;
+	}
+	50% {
+		opacity: 0.35;
+	}
+}
+
+.errorPopup {
+	max-width: 320px;
+	padding: var(--wpds-dimension-padding-md);
+	background: var(--wpds-color-bg-surface-strong, #ffffff);
+	border: 1px solid var(--wpds-color-stroke-neutral-weak);
+	border-radius: var(--wpds-dimension-radius-md);
+	box-shadow: var(--wpds-shadow-elevation-md);
+	color: var(--wpds-color-fg-content-neutral);
+	font-size: var(--wpds-font-size-sm);
+	line-height: 1.4;
+	display: flex;
+	flex-direction: column;
+	gap: var(--wpds-dimension-padding-xs);
+}
+
+.errorPopupTitle {
+	font-weight: 600;
+	color: #dc2626;
+}
+
+.errorPopupMessage {
+	white-space: pre-wrap;
+	word-break: break-word;
+}
